Simple calculator for geometry in Python by Mewwaa
